erstatt tester i KodeverkKlientTest med spock og RestClientTest
